Add ProofRule::CHAIN_M_RESOLUTION, disabled by default (#11984)
This adds the definition of a proposed addition to CPC which mimics the
behavior of what is currently called `MACRO_RESOLTUION`.

In my initial tests, the use of this rule is significantly faster than
an elaboration to chain resolution.

In a followup, we should eliminate MACRO_RESOLUTION /
MACRO_RESOLUTION_TRUST in favor of this rule to avoid redundant work in
the post processor.