Move distinct to theory UF (#12118)
Previously `distinct` belonged to builtin theory, whose theory solver
does not expect to receive constraints.

This moves `distinct` to Theory UF, to enable an approach where
`distinct` is not eliminated at preprocessing and is handled dynamically
during solving.

This does not yet change the rewriter for distinct, which will be done
in a followup PR.