Fix soundness of 5 RARE BV proof rules (#12184)
The following proof rule was unsound due to a typo in the returned term:
```
uf-bv2nat-geq-elim
```

For the others, note the term `(@bv n w)` represents a bitvector
constant of width `w` and value `n`.

In detail, `(@bv n w)` evaluates to the bitvector constant of width `w`
and value `(n mod 2^w)` when `w` is a 32 bit unsigned integer value. In
particular, this means that `w` can be `0` and `n` is allowed to
overflow.

Based on these semantics, at least 4 of our RARE BV rules are unsound.
(cvc5 will never invoke them in an unsound way, as it does not reason
about BV of 0 bitwidth currently or non-normalized bitvector constants).
In this PR, we keep these semantics and fix these 4 rules.

We add guards to the following rules to insist on positive (non-zero)
bitwidth:
```
bv-ult-one
```

We update the following rules to avoid issues related to overflow values
of BV constants:
```
bv-concat-merge-const
bv-shl-by-const-2
bv-lshr-by-const-2
```

In all cases, the rules are currently unsound as discovered by
counterexamples found by the SMT formalization of Eunoia.

Also updates the internal rewriter to be in sync with the evaluator for
`@bv`.

There are also some additional issues with the BV overflow elimination
rules which I am investigating and will address in another PR.

Note: this is an alternative to https://github.com/cvc5/cvc5/pull/12155,
which changed the semantics of `@bv`.