Do not split on infinite datatypes based on fmf option (#12218)
Previously our behavior was somewhat inconsistent: we would not split on
datatypes that had a possible constructor that was "interpreted finite"
(i.e. finite only when uninterpreted sorts are finite), even when finite
model finding is disabled (i.e. uninterpreted sorts are considered
infinite).

This updates the behavior to be more consistent: we do not need to split
on datatype terms that have a constructor that has infinite cardinality
based on the option.

This change leads to noticeably better performance on the verus set
(4.7% faster, +32-7 over a 60 second timeout).