Copy utility methods to theory model for delayed function models (#12230)
This is in preparation for faster model construction, where the models
for uninterpreted functions are constructed on demand in TheoryModel
instead of eagerly in TheoryModelBuilder.

This copies over the two key methods for model construction (first-order
and higher-order) and makes them private to TheoryModel so they can be
called on demand.

It also makes minor modifications to allow `const` to be given to the
methods.