Fix double negation handling in NNF proofs (#12276)
Fixes https://github.com/cvc5/cvc5-projects/issues/773.

Double negation was handled lazily in NNF by deferring further rewriting
to another call to NNF.

This policy could lead to inconsistent proofs if there were further NNF
steps that were processed in the current rewrite and within the formula
that was deferred.

This ensures we process NNF on the doubly negated formula in the same
rewrite step, thus correcting the inconsistency in proof reconstruction.